Make idempotent, follow advice & fix alternatives
Shell is not idempotent by default, so I made it report no changes.

Ansible reported the advice to replace `items:` with `name: ['…']`. Since we don't really have multiple packages, I replaced it with `name=…`.

Alternatives requires the link argument on debian systems if the alternative is not known. Fixed it.