Give up and make python3 default if in doubt
If there's no `python`, `python3` will be used.

Originally, I wanted to avoid this, because I didn't want to risk to break something.
But after quite a few problems with python2 being the default, I've given up.

This approach at least reveals and proves that the phpstorm script works with python3 as well (if so).